+++ lib/scout/command.rb
@@ -87,9 +87,9 @@ module Scout
                  "Optional https proxy for SSL traffic." ) do |https_proxy|
           options[:https_proxy] = https_proxy
         end
-        opts.on("--fqdn FQDN", String,
-                "Optional fully qualified domain name override." ) do |fqdn|
-          options[:fqdn] = fqdn
+        opts.on("--hostname HOSTNAME", String,
+                "Optionally override the hostname." ) do |hostname|
+          options[:hostname] = hostname
         end
 
         opts.separator " "
@@ -177,8 +177,7 @@ module Scout
       @server_name    = options[:server_name]
       @http_proxy     = options[:http_proxy] || ""
       @https_proxy    = options[:https_proxy] || ""
-      @fqdn           = (options[:fqdn] || `hostname -f`).chomp
-
+      @hostname       = options[:hostname] || Socket.gethostname
       @options = options
       @args    = args
 
@@ -190,7 +189,7 @@ module Scout
 
     end
 
-    attr_reader :server, :history, :config_dir, :log_path, :server_name, :fqdn
+    attr_reader :server, :history, :config_dir, :log_path, :server_name, :hostname
 
 
     def verbose?