+++ lib/scout/server.rb
@@ -27,7 +27,7 @@ module Scout
     attr_reader :client_key
 
     # Creates a new Scout Server connection.
-    def initialize(server, client_key, history_file, logger = nil, server_name=nil, http_proxy='', https_proxy='', roles='', fqdn = nil)
+    def initialize(server, client_key, history_file, logger = nil, server_name=nil, http_proxy='', https_proxy='', roles='', hostname = nil)
       @server       = server
       @client_key   = client_key
       @history_file = history_file
@@ -37,7 +37,7 @@ module Scout
       @http_proxy   = http_proxy
       @https_proxy  = https_proxy
       @roles        = roles || ''
-      @fqdn         = fqdn
+      @hostname     = hostname
       @plugin_plan  = []
       @plugins_with_signature_errors = []
       @directives   = {} # take_snapshots, interval, sleep_interval
@@ -49,7 +49,6 @@ module Scout
       @history_tmp_file = history_file+'.tmp'
       @plugin_config = load_plugin_configs(@plugin_config_path)
       @data_file = Scout::DataFile.new(@history_file,@logger)
-
       # the block is only passed for install and test, since we split plan retrieval outside the lockfile for run
       if block_given?
         load_history
@@ -61,7 +60,7 @@ module Scout
     def refresh?
       return true if !ping_key or account_public_key_changed? # fetch the plan again if the account key is modified/created
 
-      url=URI.join( @server.sub("https://","http://"), "/clients/#{ping_key}/ping.scout?roles=#{@roles}&fqdn=#{URI.encode(@fqdn)}")
+      url=URI.join( @server.sub("https://","http://"), "/clients/#{ping_key}/ping.scout?roles=#{@roles}&hostname=#{URI.encode(@hostname)}")
 
       headers = {"x-scout-tty" => ($stdin.tty? ? 'true' : 'false')}
       if @history["plan_last_modified"] and @history["old_plugins"]